fix: rename from option-I to option-one
This is technically a bug in pollination registry but let's fix it as is for now.
