+++ .github/workflows/ci.yaml
@@ -54,7 +54,7 @@ jobs:
       - name: deploy to staging
         run: |
           queenbee
-          pollination dsl push pollination-leed-daylight-option-I --tag ${{needs.deploy.outputs.tag}} -e https://api.staging.pollination.cloud -src https://api.staging.pollination.cloud/registries --push-dependencies
+          pollination dsl push pollination-leed-daylight-option-one --tag ${{needs.deploy.outputs.tag}} -e https://api.staging.pollination.cloud -src https://api.staging.pollination.cloud/registries --push-dependencies
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           QB_POLLINATION_TOKEN: ${{ secrets.POLLINATION_STAGING_LADYBUGBOT_TOKEN }}
@@ -79,14 +79,14 @@ jobs:
           echo "::set-output name=tag::$TAG"
       - name: install python dependencies
         run: |
-          pip install pollination-leed-daylight-option-I
+          pip install pollination-leed-daylight-option-one
           pip install -r requirements.txt
       - name: deploy to production
         run: |
           TAG=$(echo "${TAG:?}" | sed 's/[[:space:]]//g')
           echo $TAG
           queenbee
-          pollination dsl push pollination-leed-daylight-option-I --tag $TAG -e https://api.pollination.cloud -src https://api.pollination.cloud/registries --push-dependencies
+          pollination dsl push pollination-leed-daylight-option-one --tag $TAG -e https://api.pollination.cloud -src https://api.pollination.cloud/registries --push-dependencies
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           QB_POLLINATION_TOKEN: ${{ secrets.POLLINATION_LADYBUGBOT_TOKEN }}
