+++ README.md
@@ -1,4 +1,4 @@
-# leed-daylight-option-I
+# leed-daylight-option-one
 LEED daylight recipe option I
 
 This recipe uses the [annual-daylight](https://github.com/pollination/annual-daylight) recipe.
